+++ packages/models/lib/src/mappers/index.dart
@@ -10,3 +10,4 @@ export 'variable.dart';
 export 'user_info.dart';
 export 'branch_action_log.dart';
 export 'branch_json_log.dart';
+export 'branch_json_file.dart';
