+++ packages/models/lib/src/models/index.dart
@@ -17,3 +17,4 @@ export 'user_info.dart';
 export 'hex_color.dart';
 export 'branch_action_log.dart';
 export 'branch_json_log.dart';
+export 'branch_json_file.dart';
